repo_update: Add the ability to sync updated Clang builds for kernel
We are only supporting Clang 6.0+ for kernel builds to avoid unnecessary
hacks; unfortunately, these are only available on the master branch.

This is gated by a flag to ensure those who don't want to build the
kernel with Clang are not penalized in terms of bandwidth (the prebuilts
folders are shallow synced and we need to unshallow them for these).

UPDATE_CLANG=TRUE ./repo_update.sh

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>